win: snapshot.rs c_ulong→u64 widen for line_col_to_byte_offset

+++ src/runtime/test_runner/snapshot.rs
@@ -501,10 +501,10 @@ impl<'a> Snapshots<'a> {
                 );
                 let Some(byte_offset_add) = logger::Source::line_col_to_byte_offset(
                     &file_text[last_byte..],
-                    last_line,
-                    last_col,
-                    ils.line,
-                    ils.col,
+                    u64::from(last_line),
+                    u64::from(last_col),
+                    u64::from(ils.line),
+                    u64::from(ils.col),
                 ) else {
                     bun_core::scoped_log!(inline_snapshot, "-> Could not find byte");
                     log.add_error_fmt(
